27/07/2024 V19:16 Demian

+++ routes/web.php
@@ -36,33 +36,32 @@ Route::get('/carreras', [App\Http\Controllers\IDYHomeController::class, 'showcar
 
 //estos son rutas para llas citas 
 Route::get('appointments', [AppointmentController::class, 'index'])->name('appointments.index');
-Route::get('appointments/create', [AppointmentController::class, 'create'])->name('appointments.create');
-Route::post('appointments', [AppointmentController::class, 'store'])->name('appointments.store');
-Route::get('appointments/show', [AppointmentController::class, 'show'])->name('appointments.show');
-Route::get('appointments/edit', [AppointmentController::class, 'edit'])->name('appointments.edit');
-Route::put('appointments/update', [AppointmentController::class, 'update'])->name('appointments.update');
-Route::delete('appointments/destroy', [AppointmentController::class, 'destroy'])->name('appointments.destroy');
-
+Route::get('appointments/create', [AppointmentController::class, 'create'])->middleware('can:appointments.create')->name('appointments.create');
+Route::post('appointments', [AppointmentController::class, 'store'])->middleware('can:appointments.store')->name('appointments.store');
+Route::get('appointments/show', [AppointmentController::class, 'show'])->middleware('can:appointments.show')->name('appointments.show');
+Route::get('appointments/edit', [AppointmentController::class, 'edit'])->middleware('can:appointments.edit')->name('appointments.edit');
+Route::put('appointments/update', [AppointmentController::class, 'update'])->middleware('can:appointments.update')->name('appointments.update');
+Route::delete('appointments/destroy', [AppointmentController::class, 'destroy'])->middleware('can:appointments.destroy')->name('appointments.destroy');
 Route::get('/admin/users', [App\Http\Controllers\UserController::class, 'index'])->name('admin.users.index');
 Route::put('/admin/users/{user}', [App\Http\Controllers\UserController::class, 'update'])->name('admin.users.update');
 
 
 // Estas son rutas para pacientes
-Route::get('/pacientes', [App\Http\Controllers\PacientesController::class, 'index'])->name('pacientes');
-Route::get('/pacientes/crear', [App\Http\Controllers\PacientesController::class, 'create'])->name('pacientes.create');
-Route::post('/pacientes', [App\Http\Controllers\PacientesController::class, 'store'])->name('pacientes.store');
-Route::get('/pacientes/editar', [App\Http\Controllers\PacientesController::class, 'edit'])->name('pacientes.edit');
-Route::put('/pacientes/actualizar', [App\Http\Controllers\PacientesController::class, 'update'])->name('pacientes.update');
-Route::delete('/pacientes/eliminar/{id}', [App\Http\Controllers\PacientesController::class, 'destroy'])->name('pacientes.destroy');
+Route::get('/pacientes', [App\Http\Controllers\PacientesController::class, 'index'])->middleware('can:pacientes')->name('pacientes');
+Route::get('/pacientes/crear', [App\Http\Controllers\PacientesController::class, 'create'])->middleware('can:pacientes.create')->name('pacientes.create');
+Route::post('/pacientes', [App\Http\Controllers\PacientesController::class, 'store'])->middleware('can:pacientes.store')->name('pacientes.store');
+Route::get('/pacientes/editar', [App\Http\Controllers\PacientesController::class, 'edit'])->middleware('can:pacientes.edit')->name('pacientes.edit');
+Route::put('/pacientes/actualizar', [App\Http\Controllers\PacientesController::class, 'update'])->middleware('can:pacientes.update')->name('pacientes.update');
+Route::delete('/pacientes/eliminar/{id}', [App\Http\Controllers\PacientesController::class, 'destroy'])->middleware('can:pacientes.destroy')->name('pacientes.destroy');
 
 
 
 //Estas son rutas para admin
-Route::get('/admin', [App\Http\Controllers\AdminController::class, 'index'])->name('admin');
-Route::get('/admin/users', [App\Http\Controllers\AdminController::class, 'listUsers'])->name('admin.users');
-Route::post('/users', [App\Http\Controllers\UserController::class, 'store'])->name('users.store');
-Route::put('/users/{user}', [App\Http\Controllers\UserController::class, 'update'])->name('users.update');
-Route::delete('/users/{user}', [App\Http\Controllers\UserController::class, 'destroy'])->name('users.destroy');
+Route::get('/admin', [App\Http\Controllers\AdminController::class, 'index'])->middleware('can:admin')->name('admin');
+Route::get('/admin/users', [App\Http\Controllers\AdminController::class, 'listUsers'])->middleware('can:admin.users')->name('admin.users');
+Route::post('/users', [App\Http\Controllers\UserController::class, 'store'])->middleware('can:admin.store')->name('users.store');
+Route::put('/users/{user}', [App\Http\Controllers\UserController::class, 'update'])->middleware('can:admin.update')->name('users.update');
+Route::delete('/users/{user}', [App\Http\Controllers\UserController::class, 'destroy'])->middleware('can:admin.destroy')->name('users.destroy');
 
 
 Route::get('/bitacora', [App\Http\Controllers\BitacoraController::class, 'index'])->name('bitacora');
